SDL sound functional

+++ Core/burner/cocoa/FXRunLoop.mm
@@ -192,8 +192,8 @@ static int cocoaGetNextSound(int draw);
 	UInt32 ticks;
 	struct timeval now;
 	gettimeofday(&now, NULL);
-	ticks = (now.tv_sec - self->start.tv_sec) * 1000 +
-        (now.tv_usec - self->start.tv_usec) / 1000;
+	ticks = (UInt32)(now.tv_sec - self->start.tv_sec) * 1000 +
+        (UInt32)(now.tv_usec - self->start.tv_usec) / 1000;
     
     return ticks;
 }
